Switch to MIT license
I’ve spent more time than I’d like to admit thinking about this. :) Started to
go in circles and also, a bit crazy (I wish I was kidding :)). This should have
been done a while ago. Was going to make it LGPL, as it is a library, but it
still makes things simply too complicated (or GPLv2.0 with special linking
exception, like libgit2). This is a small library. Want folks to use it easily.

Enjoy.