Handle sign bit for SP78 data type
Think this addresses #15. So it's now -1 degree Celsius. My theory on this is
that certain sensors at lower temperatures have inaccurate readings. Maybe there
located close to an internal fan, and the airflow is causing it to be skewed? In
any case, this at least lets us feel good that the machine is not in danger of
over heating.

https://github.com/beltex/SMCKit/issues/15